requirements.txt
streamlit>=1.37.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.2
plotly>=5.15.0
Pillow>=10.0.0
google-generativeai>=0.7.2
